requirements.txt
pillow>=9.0.0
pytesseract>=0.3.9
